Expose an Internal module
