+++ slack-web.cabal
@@ -103,6 +103,7 @@ library
       Web.Slack.Classy
       Web.Slack.Common
       Web.Slack.Conversation
+      Web.Slack.Internal
       Web.Slack.MessageParser
       Web.Slack.Pager
       Web.Slack.Types
@@ -124,9 +125,9 @@ library
     , http-api-data >= 0.3 && < 0.6
     , http-client >= 0.5 && < 0.8
     , http-client-tls >= 0.3 && < 0.4
-    , servant >= 0.12 && < 0.20
-    , servant-client >= 0.12 && < 0.20
-    , servant-client-core >= 0.12 && < 0.20
+    , servant >= 0.16 && < 0.20
+    , servant-client >= 0.16 && < 0.20
+    , servant-client-core >= 0.16 && < 0.20
     , text (>= 1.2 && < 1.3) || (>= 2.0 && < 2.1)
     , transformers
     , mtl
